Moves images related functions in gulp/images.js
